requirements.txt
cuda-pathfinder==1.5.2
cupy-cuda12x==14.0.1
iniconfig==2.3.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
six==1.17.0
tqdm==4.67.3
